requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
amqp==5.2.0
attrs==24.2.0
billiard==3.6.4.0
blinker==1.8.2
celery==5.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.1
multidict==6.1.0
numpy==2.1.2
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.14.0
